pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "newspaper-explorer"
version = "0.1.0"
description = "A tool for exploring and analyzing newspaper content"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "newspaper-collective"}
]
keywords = ["newspaper", "explorer", "analysis"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "requests>=2.31.0",
    "tqdm>=4.66.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "polars>=0.20.0",
    "lxml>=5.0.0",
    "pyarrow>=15.0.0",
    "duckdb>=0.9.0",
    "pydantic>=2.0.0",
    "natsort>=8.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "types-requests>=2.32.0",
    "types-tqdm>=4.67.0",
    "lxml-stubs",
]
nlp = [
    "spacy>=3.7.0",
    "germalemma>=0.1.3",
    "unidecode>=1.3.0",
    "pyphen>=0.14.0",
]
normalize = [
    "transformers>=4.30.0",
    "torch>=1.12.0",
]
analysis = [
    "openai>=1.0.0",
    "anthropic>=0.3.0",
    "gliner>=0.1.0",
    "bertopic>=0.15.0",
    "sentence-transformers>=2.2.0",
    "scikit-learn>=1.3.0",
]
ui = [
    "streamlit>=1.28.0",
    "plotly>=5.17.0",
    "pandas>=2.0.0",
    "neo4j>=5.0.0",
    "Pillow>=10.0.0",
]
images = [
    "Pillow>=10.0.0",
]
all = [
    "newspaper-explorer[nlp,normalize,analysis,ui,images]",
]

[project.scripts]
newspaper-explorer = "newspaper_explorer.main:cli"

[project.urls]
Homepage = "https://github.com/newspaper-collective/newspaper_explorer"
Repository = "https://github.com/newspaper-collective/newspaper_explorer"
Issues = "https://github.com/newspaper-collective/newspaper_explorer/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
newspaper_explorer = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "integration: marks tests as integration tests (require actual data files)",
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
namespace_packages = true
explicit_package_bases = true
mypy_path = "src"

[[tool.mypy.overrides]]
module = "newspaper_explorer.*"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = ["pyphen", "germalemma"]
ignore_missing_imports = true